Allow Ignoring Included Directories
Previously the path passed to the executable had to exactly match the
include path to meet the requirements of being "this header files
import". This was inconvenient when working in a project whose directory
structure has a base path that is always searched for headers such as
"src/" or "include/".

This commit adds the ability to ingore leading paths when determining if
the file being processed matches the current header being imported.
Right now it is not configurable and hard-coded to "src/" and "include/"